kafka - websocket 설정

+++ src/main/java/ks/com/budgetmanagementproject/feature/budget/dto/BudgetSettingRequest.java
@@ -10,6 +10,7 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.YearMonth;
 
 @Getter
@@ -20,7 +21,7 @@ public class BudgetSettingRequest {
 
     @Schema(description = "설정 예산", example = "100000")
     @NotNull(message = "예산을 입력해주세요.")
-    private long money;
+    private BigDecimal money;
 
     @Schema(description = "예산 카테고리", example = "식비")
     @NotBlank(message = "카테고리를 입력해주세요")
