+++ src/main/java/ks/com/budgetmanagementproject/feature/budget/dto/BudgetUpdateRequest.java
@@ -7,6 +7,8 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
@@ -15,6 +17,6 @@ public class BudgetUpdateRequest {
 
     @Schema(description = "설정 예산", example = "100000")
     @NotNull(message = "예산을 입력해주세요.")
-    private long money;
+    private BigDecimal money;
 
 }
