+++ src/main/java/ks/com/budgetmanagementproject/feature/expenditure/dto/ExpenditureDetailResponse.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Getter
@@ -19,5 +20,5 @@ public class ExpenditureDetailResponse {
 
     private boolean excludingTotal;
 
-    private long money;
+    private BigDecimal money;
 }
