+++ src/main/java/ks/com/budgetmanagementproject/feature/expenditure/dto/ExpenditureGuideResponse.java
@@ -5,6 +5,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Getter
@@ -15,5 +16,5 @@ public class ExpenditureGuideResponse {
 
     private List<ExpenditureGuide> guideList;
 
-    private long totalAmount;
+    private BigDecimal totalAmount;
 }
