+++ src/main/java/ks/com/budgetmanagementproject/feature/expenditure/dto/ExpenditureList.java
@@ -5,13 +5,13 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
 public class ExpenditureList {
-
     private String memo;
 
     private LocalDate period;
@@ -20,5 +20,5 @@ public class ExpenditureList {
 
     private boolean excludingTotal;
 
-    private long money;
+    private BigDecimal money;
 }
